pyproject.toml
[project]
name = "vectoriseontop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "opencv-python>=4.9",
    "pydantic>=2.12.5",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "python-multipart>=0.0.6",
]
